Add HumanGeomData.h containing processed model from Human Shapes.
This anthropometric model is based on the model from Human Shapes
(http://humanshape.org/), used with permission of Prof. Matthew P. Reed,
University of Michigan.  Prof. Reed provided the data files and example
Python code to guide this work.

To speed execution and minimize memory use, the left-hand side of the
mesh was discarded (a symmetrical mirror will be inserted later).  Then
the mesh was coarsened to 20% of the original number of faces using
Matlab's 'reducepatch'.  This routine does not move, add, or delete
any vertices -- it re-connects existing vertices into a coarser mesh
while attempting to preserve features like edges.

Since original points were used, the principal component fitting
coefficients for those verts can continue to be used without need for
interpolation or blending.

The skeleton joint locations were taken from landmarks included in the
same data set -- with the exception of the elbow joint.  The elbow joint
was created as the midpoint between the elbow landmark and a mesh point
in the crook of the elbow.  The principal component coefficients were
likewise averaged to synthesize values for the new joint location.

The skeleton will be used with the Pinocchio auto-rigging tool
(https://github.com/elrond79/Pinocchio) to make this model posable.