OpenVSP 3.18.0
One major feature, lots of tweaks, quite a few fixes.  This release comes
a bit before the Workshop (see you then).  Expect another release in the
reasonably near future with a bunch more features in the works from
ESAero.

In this release, OpenVSP finally gets a built-in human figure to act as
pilot, passenger, or whatever you need.  Although it isn't perfect, it
should be better than past attempts (sorry PodMan and Male90Pct) in just
about every possibble way.  Note that this component is implemented as
a mesh (not a Bezier surface), so you can't CFDMesh, FEAMesh, or export
it as STEP, IGES, or a few others.

Lots of user-requested tweaks -- more Sets (20 for now), different color
highlights for XSec editing, (near) eradication of non-adjustable sliders.

All good stuff -- but the reason to update is to improve the pilot/pax in
your models.  You know you want to.

Features:
 - New anthropomorphic pose-able human model
 - More sets by default (20)
 - Blue and green highlights for wing/fuse/stack editing
 - Convert non-adjustable sliders to adjustable
 - Improved Cpslice visualiation
 - Other small cleanups

Fixes:
 - Fix some API Error messages
 - Don't draw Prop XSecs when in disk mode
 - Fix memory leak