Fully Defines the Generic XSec Class (EditCurveXSec)
Based off of the PCurve class, this new XSec type defines a series of
functions for advanced point control of an XSec's Bezier Spline. The
class is based on the PCurve class, but has a number of different
features. The curve is defined by X and Y 2D coordinates and a U
paramer (parm vectors). Methods include initialization of the XSec
as a circle, oval, or rectangle, curve splitting, moving and deleting
points, conversion between CEDIT, PCHIP, and linear,G1 continuity
enforcement for CEDIT, and Left/Right Symmetry enforcement. For
more details, see the comments in XSecCurve.h.