Adds Functions to Convert Any XSec Type to Edit
The XSecSurf version finds the XSecCurve specified by the input XSec
index, calls ToBinaryCubic on the baseline curve saved in XSecCurve::
Update (reduces the number of control points for XSecs with an
excessively large amount, like airfoils), converts the XSec to
EditCurveXSec type, and copies over the width and height. The BOR
version is required, since BOR Geoms are not GeomXSec.