Adds API Functions for Edit XSec Types
These functions should provide all of the capabilities available
through the GUI for EditCurveXSec types. They include functions
to initialize the XSec to the value of m_ShapeType, convert to an
input type (i.e. linear), get the  curve parameterization vec, get the
control points (dimensional or not), set the U and control points,
split the curve, move or delete a point, and convert an XSec to
EditCurveXSec type (works for BOR with XSec index ignored).