Extend Each Register Call to Allow Addition of a Comment
First, an "asDocInfo" struct is defined to specify a comment and
associated group. This struct can be passed into the registration
functions, which add the comments and groups to the asDocGen comment
and group maps. "m_ObjTypeMap" is used to associate member and function
names after they've been reformatted by the AngelScript registration.