Fix ordering of some variables to DegenGeom CSV file.
Some variables were added to the middle of a variable list rather than
strictly appending to the end.  This broke VSPAERO and possibly other
codes.

Unfortunately, the mistake was released in 3.16.2 (Oct 2018), so it is
possible that other tools that use the DegenGeom CSV file have come to
expect this behavior.