OpenVSP 3.19.1
Right on the heels of .0, 3.19.1 is mostly about tidying up a few loose
ends.  There are a few bug fixes and some documentation improvements.

Most significantly, if you're interested in using VSPAERO to analyze a
case with rotating blades, this version fixes an issue with DegenGeom
and also includes some example cases for you to try.

Not much for everyone else, but no reason not to have the latest either.

Features:
 - Add rotating blade examples for VSPAERO
 - Let vspviewer accept *.adb in file name passed, helps with double-click
 - Use less precision in M,A,B list from GUI to VSPAERO -- long command lines
 - Clean up some warnings in VSPAERO code
 - Update VSPAERO command line documentation

Bug Fixes:
 - Fix field ordering in DegenGeom file -- this broke VSPAERO
 - Fix problem with image backgrounds
 - Remove XS_BEZIER reference from Python test
 - Improve Ubuntu package to remove prior version
 - Fix PathToExe for FreeBSD