OpenVSP 3.20.0
OpenVSP 3.20 just in time for 2020!  Start off the new year with a new
version with one really cool feature (Smart Input) and a bunch of fixes.

Overall, this update is very small.  It wouldn't cause a .20 bump except
that the new feature also adds a new library dependency and that warrants
a little more attention.

You may be asking, "What is Smart Input?".  Input boxes (the white box
with parameter values) are now smarter.  They can now can do simple math
(+-*/^).

Maybe you want to nudge a parameter by 0.375 -- just click in the input
box and put '+.375' after the current value.  Maybe you want to double
something '*2', increase by 10% '*1.1', convert units on the fly '/2.54'.
This is a really handy feature that will find many uses.

In addition to Smart Input, there are a handful of build fixes, code
cleanups and improvements to the automation of the build process.

Features:
 - Smart Input for simple math in input boxes
 - API examples in documentation
 - Draw cg from mass prop screen
 - CEDIT improvements
 - Make wing section projected area an output Parm
 - Add get parent/child to API
 - Fix VSPAERO crash for Uber CRM example case
 - Lots of cleanups identified through static analysis
 - Travis deployment
 - Update Travis MacOS build to match release builds

Library Updates:
 - Include expparse simple expression parsing library

Bug Fixes:
 - Don't crash structures STEP/IGES output when no structure defined
 - Parasite drag for identically named geoms
 - DegenGeom crash when subsurfaces extended beyond [0,1]
 - CEDIT fixes