Transfer Working Link Settings to New Added Link
Previously, adding a new link would cause all link parameters to be
reset to their defaults. The user would then have to click on the new
link in the browser to set it to the working link and set the parameters
from there. This commit automatically transfers the parms from the
working link to the new link. (Identified by Andy Hahn)