Correct calculation of Schemensky's FR for body form factor.
Schemensky's FR=len/sqrt(w*h) rather than len/dia.

The prior calculation was also wrong in many ways.  It attempted to
calculate len/sqrt(Area) (wrong), but instead of max cross sectional
area (sectArea), it used the wetted area of the 'top' of that part of
the section (areaTop).

This change uses the chord as the section width and the toc*chord as the
section height.  toc is valid for all XSec types (not just airfoils)
whereas toc2 likely only works for airfoils.