BUG FIX: CEDIT Aspect Ratios Must be Set at the End of DecodeXML
Because of the order of the width, height, and aspect ratio parms in the
XML file (decoded alphabetically), the aspect ratio would be calculated
before width was decoded. The aspect ratio is now calculated after
all parms are decoded for the XSecCurve. In addition, the unecessary two
versions of aspect ratio are replaced with one, since they were the
inverse of each other previously.