Adds & Registers Separate XSec Height and Width Setting Functions
This is mainly done because if aspect ratio is preserved for a CEDIT
XSec, the input width in SetXSecWidthHeight is ignored. Height will
be used to set the width based on aspect ratio. These functions allow
height to set width or width to set height when AR is preserved. Note,
The SetWidthHeight function for CEDIT is also updated to be consistent
with other XSecCurves.