BUG FIX: Prevent colorChooser Update When Entering Numbers
This call is unecessary and was causing a max of one digit to be entered
in the colorChooser inputs before they were reset from the update. The
callback is sufficient to set the values and update the background
color. In addition, setMeshColor is called directly instead of
setRed, setGreen, and setBlue for greater efficiency.