Dynamic Resizing for Parm Link Browser Columns
This allows the columns of the browser to be resized when the window is
resized and then updated. A int pointer is defined to specify the
array of widths so that the values can be updated without changing the
memory address. Note, deallocating the width pointer will cause an
exception to be raised.