VSPAERO Update 6.0.2
1. Added time histories of blade loading for the last revolution, depending
on if it's a single rotor, all rotors at the same RPM, or if all are are different.
2. Renames P, Q, & R output files from .stab to .pstab, .qstab, and .rstab
3. Added logic  so view does not die, fix in in solver not viewer
4. -noise option for psuwopwop now creates multiple namelist files:
entire vehicle + rotors, just the vehicle w/o rotors, and a name list for
each rotor + the vehicle. The case.nam file has the full configuration set
to run, you can uncomment all or some of the other cases. The impact
of each rotor individually on the noise can be assessed. Note, these are
from the same vspaero run, so you get all the aero interactions, but
psuwopwop does not evaluate the sound for all of them.
5. Solver now splits components into separate groups per symmetry, including
wings. Any old .groups files for models with wings, htails... will be broken.
6. Added in the wake iter 0 logic to do no wake iterations