Make smart edit conditional part of build via NOREGEXP
std::regexp was introduced in gcc 4.8 as a broken stub and really doesn't
work until 4.9.  Some important computers are stuck on really old
compilers, so this change helps them build and run at the expense of
the smart edit functionality.

These computers are better off building with VSP_NO_GRAPHICS in most cases.