Adds a Fixed Wake Option for VSPAERO (Includes API and GUI Support)
This flag sets the number of wake iterations to 0, indicating a fixed
wake in VSPAERO. Support is added to the Analysis Mgr and VSPAERO
Screen. Note, this commit also resizes the sliders in the Wake section
to align with the number of CPU slider. The start and end wake iteration
sliders are deactivated for fixed wake and time dependent analysis.