BUG FIX (872a6d3): Fix Broken Pressure Conversions in Parasite Drag GUI
In 872a6d3d611568bebafda0823414144e1c88e766, enums for pressure and
density were added to the middle of the enum set. The change to the
other density enums did not change anything, but the pressure enums
caused the dropdown in the Parasite Drag Screen to be off. They are
added as additional dropdowns, but no change to conversions is needed,
since it has already been implemented. Support is also added for
exporting the pressure unit names.