Changes X & Y Parms for Generic XSec to FractionParms for Scaled Display
The X and Y Parm vectors for Generic XSecs are changed to FractionParms.
This allows the reference value to be set to the XSec width and height
for X and Y positions respectively when the XSecCurveEditor is
updated. If the flag to view absolute values of the cross section is
true, the FractionParm results are shown.