Improves Identification of Intersection Points on Patch Edges
A planar approximation is used to find the closesest UW values for each
intersection point with respect to each paranet surface. The check for
interseciton points on the edge of a patch is then performed on these
UW values with a tighter tolerance. The planar approximation is used as
an initial guess for the final UW identification. Overall, this change
does not influence the intersection algorithm's speed significantly.