Update VSPAERO to 6.0.3 from Dave Kinney
Fixed the centerline issue for cases with symmetry OFF - this was caused
by the splitting of wings across two components in the last version to
make symmetric rotors work... basically it broke code that is looking
for near overlapping components. So thinks like vtails overlaying the
cruciform body where tweaked so that the code would converge and the
solutions were reasonable. But this relied on a check that assumed a
compoent would not intersect itself... but a split wing does just that.
So, lots of new bbox checks that make the whole process more general
and clean anyway.

Fixed the side slip at the symmetry plane jump that Rob noted. This
was a book keeping error in the loads, force calcs.