Updates Load Tab of VSPAEROPlotScreen to Support Average Blade Load Dist
The results in the rotor files generated from VSPAERO unsteady analysis
include average load distribution fro each blade in the rotor. The
implementation of plotting on the Load tab is similar to the
implementation of other unsteady result plotting on the Unsteady tab.
A Parm is also added to VSPAEROMgr with associated enum for
identifying if load or blade distribution is being plotted