Improves VSPAERO Panel Method Mesh Treatment
If a mesh is the only geom in the set, another is not generated. This
allows meshes generated outside of VSPAERO to be run. If there is more
than one geom in the set, a new mesh is created. Overall, this
significantly improves the behavior of generating meshes for VSPAERO
panel method.