Updates Triangle Intersection Algorithm with ISeg Usage
This version of the triangle intersection with intersection line
algorithm is more robust than the existing one created by Thomas
Moller. However, it's implementation is very similar. Note, the
coplanar flag must be initialized to 0 before calling the function
because the function only sets it to 1 if coplanar, but not to 0 if
not coplanar.