BUGFIX: Always Reset Curve Editor Slider Highlight in Update
Previously, the last selected index was being used to reset
the highlighting of the sliders if they were deselected.
Now, the slider label color is always reset in update, other
than for the currently selected point, which is highlighted
yellow This prevents a memory leak that was being
introduced when the number of sliders in the vector would
change.