Includes Component Section of VSPAERO Load File in ReadLoadFile
A flag is added to indicate which sections of the *.lod file has been
read. After the first section is read, a new result, "VSPAERO_Comp_Load"
is created and the component data is read in. "VSPAERO_Load" results
are not changed by this commit, and the new component results are not
supported in the VSPAERO Plot Sreen.