Adds, Registers, and Documents Additional FEA Structure API Functions
In response to a limitation of setting FEA Skin properties identified on
the Google Group, the API for structural modeling is siginficantly
expanded. Most important are the functions to get the IDs of all FEA
Structures or FEA Parts or Subsurfaces for a particular Structure. Other
functions are added to get or set values that are not Parms, such as
names and types. This commit also includes minor improvements to
existing FEA Structure API documentation.