Removes Unused Input to AddBorderSplit
The IPnt pointer may be used to identify the second Puw* input before
the function is called, but otherwise is not used in the function.
This also allows for IPnt pointers to not need to be created just to
provide the input. Note, "closest_proj" declared in AddBorderSplit but
isn't used is also removed.