Consolodates Checks for Invalid Surfaces and Skip Export if No Surfaces
Instead of checking the validity of each trimmed surface in each trimmed
CAD export function, it is done in BuildNURBSSurfs before each
NURBS_SUrface is added to m_NURBSSurfVec. If there are no surfaces in
m_NURBSSurfVec when the export functions are called, no files are
written and an error message is output.