BUG FIX: Set Very Small Values to 0 Before Triangle Intersections
This is done to avoid + or - assignment to very small numbers that have
experienced floating point precision losses. While small, the sign of
these numbers can cause them to not be considered in the intersection
algorithm. This issue mainly appears when the intersection point is on
a patch edge and has an X, Y, or Z coordinate of 0.