Updates Geom::SetSetFlag to Add/Remove between SET_SHOWN & SET_NOT_SHOWN
This is done to force Geoms to always be in SET_SHOWN or SET_NOT_SHOWN,
but not both at the same time. Previously, a Geom could be in both
when editing the Geom's set through the Gen tab of the Geom Screen. This
change also returns early if the index is out of bounds or if the set
flag is already set.