Fixes Endless Update Loop, Removes Unecessary Variable, and Renames Parm
And endless update loop was created in PSliceScreen when the start and
end slice location parms were set, which would trigger another update.
A check is added to see if the min and max bounding box has changed
before updating. The Planar Slice axis name parm is updated for clarity.
The index that tracks the axis in PLiceScreen is unecessary and
therefore removed.