Allow zero and negative values for Planar Start/End Location
This was causing a bizarre GUI loop when the GUI would try to set
zero or negative values -- these would get clipped by the Parm bounds,
triggering a new update, which would try to set back to zero or negative.

This solves the bizarre loop and also allows reasonable legal values.