Removes Flow Condition from VSPAERO Command Arguments
Since they are specified in the VSPAERO setup file, there is no need to
include them as command line arguments. For non-batch mode cases, the
setup file is written for each flow condition case. An index is added
to track the current flow condition case and appropriately identify
the alpha, beta, and mach values.