Adds ColResizeBrowser to GuiDevice
This special type of Fl_Browser allows for the borders of each column in
the browser to be shown and resized by the user. This is a modification
of the code provided by Erco's FLTK Cheat Page
(seriss.com/people/erco/fltk/). The primary modifications are formatting
changes to match other GuiDevices. One of the more significant changes
is an adjustment to event handling to allow selecting items but not while
a column is being resized.