Adds Auto Hub Diameter Option for Prop Geoms in Disk or Both Mode
This is done to allow a different hub diamter to be input when the
prop is in disk or both mode. If it is in blades mode, the hub
diameter is automatically set from the geometry. Disk custom
components do not have the option to set the hub diameter
automatically.