Updates Number of Unsteady Group and Actuator Disk API Functions
A forced update is added to avoid an issue where the wrong number of
unsteady groups or rotors are returned since the VSPAEROMgr is not
updated when a change is made to geometries. An additional function
is added to get the number of unsteady rotor groups. The new
function is registerd and documented. Some fixes are made to the
GetNumUnsteadyGroups documentation.