Forces Updates of Some Managers in API Update Function
This is done to ensure expected behavior when updating from the
API. In the GUI, these managers are updated through their associated
screens. The API previously had no way to update these managers,
causing them to not respond appropriately to changes in geometry.
A flag is included in the API function to control this behavior and
documented accordingly