Adds Initial Equal Arc Length Reparameterization for EDIT_CURVE XSec
A function is added to reparameterize the U values for the curve
based on teh arc length of each segment and the total arc length of
the curve. A button is added to the GUI to perform this
reparameterization. Note, this does not include the option to leave
corners at this time, which will be necessary to support.