Set Prop as Master when RPM is Changed in Uniform RPM Mode
This is done primarily to improve behavior through the API where
RPM would not change if in uniform RPM mode and the current unsteady
group index did not correspond to the Parm being currently set. An
alternative function is added to set the current unsteady group from
the unsteady group ID to support this.