BUGFIX: Tag CFD Mesh Surfaces when SubSurfaces are not Intersected
IN CFD Mesh, the "m_IntersectSubSurfs" was being used to identify
when Subtag should be called. This resulted in surfaces all being
tagged the same when the flag was false. To ignore subsurfaces but
still tag separate surfaces, the SubSurfaces are simply not
trasnfered to the mesh and intersection managers from geom_core.