Add FlipNormal and Transform methods to DegenGeom class
These methods will allow a DegenGeom to be passed to the ApplySymm
method to transform a Main DegenGeom to all the symmetrical copies.

This does not do a full transformation, it only transforms the elements
used in the visualization preview.  This should not be used in the primary
DegenGeom code path, only in the shortcut viz codepath.