Pass number of surfaces to CreateDegenGeom( vector<VspSurf>...
m_MainSurfVec always has at least one surface in it -- however,
sometimes GetNumMainSurfs() returns zero.

This is true for blank, hinge, and other zero-surface Geom types.

A better long-term fix will be to clear those situations up and allow
m_MainSurfVec to be empty.  In the meantime, this is a prudent fix.