Adds VSPAERO Path Identification to Support API Path Variations
A string is added to Vehicle to track the absolute path to the VSPAERO
executables. The path is initially set the same as the VSP executable
path, but functions are added to set the VSPAERO path and check if the
VSPAERO executables are in a particular directory. These functions are
added in preparation of API support. VSPAEROMgr is updated to use
the VSPAERO Path instead of the VSP executable path.