Background Image Setting Parameters Moved to XSecCurve Parm Container
These parameters will allow each XSec to have a different background
image and settings that will be saved with the model. This is done in
preparation of adding background image support for Edit Curve types.
ScreenBase is updated to use these Parms instead of ones defined in the
GUI class. Improvements are made to allow improved update behavior
using settings for individual XSecs instead of all.