Make wake point matching use approximate comparison
Wake edges from the top and bottom surface of the wing were sometimes
very slightly (~1e-15) different where the wake intersects another body.

This is likely a consequence of the intersection algorithm.  Using ==
as a comparison was causing reasonably identical wake edges to be
considered different, which was causing erroneous wakes to be written
out to the vspgeom file.