Simplify overridden versions of UpdateTesselate and UpdateSplitTesselate
Also organize calls -- document reasoning and call sites, etc.

This commit is a bit of a mess, but three methods are entirely eliminated.

Only one version of UpdateSplitTesselate remains.  The three versions
of UpdateTesselate are relatively simple (could be reduced to two easily).
Two are legacy wrappers that have been around a long time.  The third
is the recently introduced version that takes a surf_vec as an argument.