Adds MATLAB API Support with CMAKE Flag
This provides a complete template for users to compile the MATLAB API
from the OpenVSP source code if SWIG and MATLAB are found. Note, the
MATLAB API requires an unreleased version of SWIG to be compiled first
and a valid MATLAB license. Because if these requirements, an additional
CMake boolean flag (VSP_ENABLE_MATLAB_AP) is added so the matlab_api
won't be included in the build unless the user sets the flag to true. The
Python API CMakeLists is also updated to check for SWIG in the parent
directory. Examples and tests are included with the MATLAB API to get
users up and running much more easily.