Adds Functions to Get Texture ID from TextureRepo and Background
This will be used to determine if a background image should be updated
if the texture IDs do not match. This will allow separate background
image files and settings to be associated with each XSec. Note, the
existing _texID in Background is not the same as the texture ID in
Texture2D.