Enhance the Wedge airfoil type.
This allows a flat spot on the upper and lower surface as well as
independent control of the upper and lower thickness profiles.  Camber
is now also supported.

These changes necessicitate the option of manual control of the U parameter
around the airfoil.  This will enable the user to control how a wing
with this sort of section lofts.