Fix use of CMAKE_SOURCE_DIR in API CMake stuff
OpenVSP can be built starting from the source root, or from the /src
directory.  Consequently, coding a path with the /src in it will not
always be correct.

Instead of using an absolute path starting from CMAKE_SOURCE_DIR, it
is better to either use relative paths ../ or CMAKE_CURRENT_SOURCE_DIR
when referring to the current directory.