Fix Update problem with clustering
Clustering parameters were not getting communicated in the case of
m_TessDirty.  This moves that code to UpdatePreTess() as well as
consolidates similar earlier work building m_UTessVec.

While we're at it, fix the the update of mesh metrics for wings and props.

Also, clustering was not being communicated to body of revolution
foil surf -- so airfoils exported from BOR would not be clustered.