Improve some API documentation
We should try to use canonical variable names in the API examples.  In this
case that means use var names that match the argument list and docs.

In cases where multiple similar calls are made this may not be possible,
but in many cases it will help to reinforce how variables are used.

This commit also fixes an error with some of the examples.  The calls
to CompNorm01 often took gid (which does not exist) instead of pid.  These
would fail to compile and run if that was attempted.