Add toint() integer conversion to Util.h
Elsewhere, we have used the C idiom (int)(x+0.5) as a conversion from
double to int.  This results in the nearest int to a double for positive
values.  Casting to int is truncation. The +0.5 moves you away from
boundary cases that can make life complicated.

Unfortunately, this idiom only works for positive values.  Instead, this
new toint() is a slightly more complex version that also works as expected
for negative values.