Adds Check for Valid DegenGeom Pointer Instead of Clearing Vector
The clearing of the DegenGeom vector after VSPAERO execution did
fix the issue of UpdateRotorDisks using a pointer of a Geom that
had been deleted, yet also caused API functions to break. If VSPAERO
had already been executed through the API, it would not execute
again unless Compute Geometry was executed again with it. This fixes
the broken API tests and examples while still preventing a crash
using an invalid Geom pointer.