BUGFIX: Uniform RPM Setting Causing Multiple Updates
in ParmChanged, the current unsteady group index is updated to the
one which RPM is set. This was done for all prop RPMs to update
correctly from the API regardless of which prop is the current one.
However, when Update was called and each RPM was set, ParmChanged
would be triggered again, causing the wrong prop to become the
current one. If the type of Parm change is SET, the current prop
is not updated.