Updates to CMake 2.8 Compatible Branch of Triangle
While OpenVSP's CMakeList expressed a minimum version of 2.8 for
CMake, the integrated Triangle library required 3.6. This modified
version of Triangle (https://github.com/justingravett/Triangle/tree/
JGravett/CMakeVersion) support CMake 2.8 or greater.
Updates Triangle Again to No Longer Need SOURCE_SUBDIR Entry

The SOURCE_SUBDIR is not available in CMake 2.8, so it is removed.
Instead, a simple CMakeLists file is added to the root directory of
Triangle to add the "src" directory.