Adds Generic XSec Point Color Option
This was requested by Kal Weller, arguing that often times background
images are black and white. This adds a toggle to make the XSec points
(on the curve) not black. A color wheel from 0 to 360 is then used to
set the color of the point, avoiding the need for a full 3D color picker