Hacky workaround for resizable column crashes in FitModel.  NOT A REAL FIX.
This adds the zero termination and changes the pointer to a static array
which then does not need to be deleted.

A better fix will be to handle the memory inside where the resizing code
is handled.

Lots of call sites need to be updated with the eventual real fix.  This
hack was just enough to let me test FitModel.