Saves Browser Horizontal Scroll Position when Updating
If the browser is wide enough to require a horizontal scroll bar, this
prevents the scroll bar from being set back to the starting position
each time an Update is performed. Updates are caused by actions such as
selecting a new item in the browser, but not from resizing columns.