Fixes and Improves Column Width Declaration for ColResizeBrowsers
Instead of calling new and representing the column widths as a pointer,
a static array is defined and then the pointer to it is passed to
GroupLayout. Zero-termination is added to all input width arrays.
Casting from double to int in the static int declatation is not supported,
so some values are changed to integers.