BUG FIX: FLTK vs OpenGL Sizes with Latest FLTK Library in Adjust View
The definition of pixel_w and/or w was likely modified in the latest
FLTK, breaking the implementation of the Adjust View screen in OpenVSP
on Windows. "factor" uses pixels_per_unit directly instead of a int/int
division that was not being converted to a double appropriately. Math
operations are avoided where possible to prevent floating point error
accumulation.