BUGFIX: Include "Master" Rows when Computing Total CD
This is in response to a bug identified on the Google Group where
the sum of CD values displayed in the bottom right of the Parasite
Drag GUI would not add up correctly if the component grouping
feature was used. This appears to be a mistake in the current
implementation where the main rows (shown in bold) should be added
up instead of the sub-rows. This is consistent with how values
such as wetter area are added up.