Fixes API Documentation Group Specification
This passing of the asDocInfo data structure was causing the group
information to become corrupted before being written to file. This
may have been due to a memory leak or something misunderstood
about the implementation. Breaking up the structure into each of its
fields fixes the issue so that groups are written out correctly.