Fixes for ellipsoid test case.
vlm was basically failing since the solution is zero vorticity and well...
that's the initial guess. Thought I'd worked around this, but not...
so I catch it explicitly now.

panel was the load distribution mods for panel breaking for something
with no wings. Thought I already caught this, but after staring at it
I clearly did not have enough coffee when I code that...