Adds Simple Half Mesh Option for Unintersected Mesh Generation
This feature is added to support half mesh for VSPAERO fully VLM VSPGeom
input files. Triangles with -Y center coordinate are thrown away. Note
if the geometry is not entered along the XZ plane or is rotated about X
this feature will result in a jagged edge due to it's simplicity.