Modify Vehicle::WriteVSPGeomFile to not Remesh Existing DegenGeom Mesh
This allows for an existing DegenGeom MeshGeom to be used in VSPAERO.
For instance, rerunning VSPAERO in VLM with a VSPGeom input. A mesh will
first be generated, then only the mesh will be shown. This mesh can then
be used again if the VSPAERO set is SET_SHOWN since the degen_set is now
included in the set checks when writing the data to file.