Adds a Function to Read Polar File and Combine Data for ReCref Sweeps
This function reads the Polar file output from VSPAERO and generates a
new result called "VSPAERO_Polar". The ReCref value in the Polar file
is compared to the start ReCref. If they are different, data from a
History file with matching alpha, beta, and mach is copied to a new
VSPAERO_History result. CDo, CDtot, and L/D are then copied from the
Polar file to the new VSPAERO_History result. ReadHistoryFile is updated
to include the ReCref flow condition as a result. Only a single
VSPAERO_Polar is generated for multiple flow conditions to allow easy
comparison in the CSV export.