Adds Support for Plotting Multiple ReCref Conditions
For convergence and sweep plots, ReCref is added to the flow condition
string. ReCref does not impact the load or CPSlice distribution so it is
excluded in the flow condition string. While unsteady cases can only run
a single flow condition, the ReCref value is included in the flow
condition string for reference.