BUG FIX: CPSlice Results not Being Cleared
CPSlice results are cleared with all other VSPAERO results instead of
only when the CPSlice activation flag is toggled on. This prevents
results remaining in the ResultsMgr and plot screen when the toggle is
turned off after a CPSlice execution. This is consistent with other
VSPAERO results. In addition, there was a type in the CPSlicer result
name.