Add rudimentary error checking for browser column width input.
This does a reasonable job catching problems, but it is bad at telling
you where the problem occurs.  It tries to spit out the hierarchy of
group labels, but that seems to mostly be tab titles.

In the test cases it found, it only succeeded in printing out tab labels
such as 'Sub' and 'Disk'.  While that may lead you to Geom Subsurface
and VSPAERO Actuator Disk, it might also lead you astray.