Removes Swig Version Identification in GitHub Actions Build Script
There was a bug that was present in Swig 4.0.2 but has been fixed in
recent versions, allowing for the latest to be used. In addition, this
is necessary to update because GitHub actions will fail using this
version of Swig since a more recent one is already installed by default.