Don't force static system libraries for vspaero with gcc on non-apple.
On Apple, GCC and its accompanying libraries are often not installed.
So, statically linking the system libraries results in a more portable
binary.  However, on Linux it is reasonable to expect that the
dynamic system libraries are installed.  In fact, the static libraries
needed for this compiler option to work are likely to not be installed
on some machines.

So, this hopefully results in the easiest time building and running
on as many platforms as possible.