Only Allows One Parm ID per Variable Preset Group and Updates API Error
This was only implemented partially prior to this change where a drag
and drop operation in to the Parm Picker would be denied if the Parm was
already in a group but would work using the dropdown choices. In that
case, the variable presets would not work when a Parm was in multiple
groups. This avoids that issue and improves consistency of the
behavior. This also adds an error message if the AddVar function does
not work from the API.