Use FindPythonLibsNew instead of FindPythonLibs
Copy
https://github.com/pybind/pybind11/blob/master/tools/FindPythonLibsNew.cmake

as a replacement for FindPythonLibs.  This version takes a hint from
the GitHub Actions variable pythonLocation to hopefully set all other
variables appropriately.

This allows our GHA build.yml to be simplified as we do not need
to pass PYTHON_LIBRARY and PYTHON_INCLUDE_DIR to CMake.

We also do not need to use setup-python on Linux as we instead want to
rely on Ubuntu's package management system to provide Python there.

Move the IdentifyLibs debug statement lower in the file to be sure
it happens in the right place.  Should be commented out long-term, but
it is handy for testing.

Add PYTHON_VERSION to IdentifyLibs