Fixes #195: Surface ID Mispatch with Surface Vector Index Label Error
This adds a function to match a surface ID with a Surf pointer in
m_SurfVec. This allows the Surf ID to be matched with the appropriate
index in the m_SurfVec vector. The problem was caused by an assumption
that the ID and index were equivalent and did not account for dropped
surfaces such as those on a sharp trailing or leading edge.

(cherry picked from commit 3ad45fd1bdd1fc2a90878bc9e8cabefc45994c2f)