Make XSecViewScreen:Update() and GuiDeviceCallBack() use XSecCurve
These were using xs->GetXSecCurve() extensively when they really should
have just stored xsc.

In particular, this will pave the way for making them work with any
Geom that uses an XSecCurve but is not a GeomXSec.  i.e. a Body of
Revolution.

At present, clicking the Show XSec button on a BOR causes a crash
because geom_xsec is NULL, but is dereferenced anyway.