Make XSecCurve::UpdateCurve() pure virtual, eliminate concrete XSecCurve
XSec's constructor can only be called via XSecSurf::CreateXSec, which
should always succeed in creating an xscrv_ptr to pass to the constructor.

However, in the event that it does not, creating a circle instead of
a shapeless XSecCurve should be OK.

Arguably, although XSecCurve could be instantiated, it should have been
abstract -- for example, the m_Curve it would have produced would have
always been empty.