Add updateParms argument to UpdateCurve()
This argument (default true) allows us to disable updating any Parms in
the UpdateCurve() process.

XSecCurves like rounded rectangle will adjust the radius parameter if the
height and width make it impossible to achieve.

That change could adversely influence the nonlinear solution process when
area is used as a driver (instead of width and/or height).

Setting updatParms=false when UpdateCurve() is called from the solver
will prevent these updates from occurring.