Add XSecCurve::AreaNoUpdate()
This calculates the area of an XSecCurve based on the current value
of m_Curve without calling Update() to ensure the curve is up-to-date.

Most callers should call ComputeArea() which ensures a curve is up-to-date
before calculating the area.  However, in the context of the nonlinear
solver used when area is a driver, we want to calculate the area
mid-update (and obviusly without triggering an update).