Fixes Reverse Flag for Prop Geoms in Disk Mode and with Symmetry
The Reverse Normal flag is checked to flip the main surface normal
vector, similar to what is done in Prop mode. However, instead of
providing the flipped normal vector to VSPAERO, the RPM of the disk
is flipped to avoid negative thrust in VSPAERO. Also, the Flip Normal
Flag for actuator disks is saved regardless of it's value instead of
if it is True. This was causing XZ symmetric copies of the Disk Custom
Component to spin in the same direction as the main disk.