Updates Normal Vector vs RPM Flip for Prop Reverse Flag
This change results in the behavior for Props and Actuator Disks
being consistent in VSPAERO, where the reverse flag is used to
set a negative RPM but not change the sign of the nomal vector
written to the VSPAERO input files. Now, thrust is output from
VSPAERO with a positive sign instead of being negative with a
flipped normal vector.