Implement GetProjectionCosine()
This queries the XSecSurf's principal direction to determine which
axis projection is relevant.

Initially, it was expected that separate GetProjectionCosine methods would
be needed for each XSec type (fuselage, stack, etc).  However, this
single implementation should be fully general.

The main purpose of an XSec is to contain a XSecCurve and to set
m_Transform appropriately.  Obtaining the projection directly from
m_Transform should be robust.