Remove support for projected area from area parameterization support.
This was just a stub in anticipation of supporting projected area.
However, the orientation of an XSecCurve (which determines the projection
cosine) is determined by the XSec and the XSecSurf.

Consequently, when the orientation changes, there is no way to notify
the XSecCurve of the change.

If this did work, there could be significant unintended consequences
particularly for Stack where the orientation of each XSec depends on
all the previous XSecs.

It seems best to just remove this and not implement projected area
parameterization.