Improves Behavior for Deleted Variables in Advanced Links
Instead of deleting the entire Advanced Link when a Parm can no longer
be found, the variable is simply deleted and an error message is printed
to the terminal. When the script is compiled, an FL message will be
displayed to the user that the variable does not exist. This change was
suggested by AJ Field.