Allow negative MeshGeom etc.
Original code was buggy -- it did what was wanted, but not what matched
documentation or intent.  When updated to do what was intended, what was
desired was lost.

This fixes comment, makes intent clear, and does what is desired.

For the record, Even when GetNumMainSurfs() returns zero,
m_MainSurfVec.size() is usually 1.  This is because we add a empty VspSurf
in the end of the Geom() constructor.  It does not get removed for cases
like Blank, Hinge, MeshGeom, Human, PtCloud, WireGeom, etc.