Swaps WingGeom's Use of m_ActiveAirfoil vs m_ActiveXSec
Instead of using m_ActiveXSec to refer to an XSec, WingGeom was using a
separate Parm called m_ActiveAirfoil. m_ActiveXSec was being used for
the wing section. This is different from every other XSec Geom type so
a new Parm is defined to keep track of the wing section and m_ActiveXSec
is used for the XSec.