Tweak subdivision limit on ToBinaryCubic to keep refinement from going too far.
Brian German provided a test case where wings using rounded rectangle
XSecs (with the improved corners) would fail to form proper surfaces.  This
only occurred with tip caps turned on.

ToBinaryCubic was resulting in 15 subdivisions in some location on the
curves.  This resulted in a tiny curve sliver (in parametric space).

The parameter split matching process for top/bottom surfaces required
to build a cap ended up with splits that did not match up and caused
the surface to be corrupt.

Restricting subdivision depth seems the most robust way to avoid this
category of problems.